Cleanup class/module names for consistency, redundancy removal, etc Overview  * Asterisk* dropped when already under Asterisk module  * Config*Generator(s) => ConfigGenerator  * Removed module module redundancy in namespaces  * removed .conf from file names
